// src/types/variables.ts
import { RGBAValue } from './figma';

// Define types for UI state
export interface Variable {
  id: string;
  name: string;
  value: string;
  rawValue: RGBAValue | string | number | boolean | null | Record<string, unknown>;
  modeId: string;
  collectionName: string;
  isColor: boolean;
  valueType: string;
  referencedVariable?: {
    id: string;
    collection: string;
    name: string;
    finalValue: unknown;
    finalValueType: string;
  };
  description?: string;
}

// Type for the select options
export interface SelectOption {
  value: string;
  label: string;
}

// Define a type for dropdown options
export interface VariableOption {
  label: string;
  value: string;
  original: Variable | null;
  isCustom?: boolean;
  type: string;
  color?: RGBAValue;
} 